SignatureHash(): Document buggy but consensus-critical legacy behavior, aid in preventing a similar fate for BCH signature hashing (#1136)
* cashlib: Use generic SignatureHash()

Use the generic SignatureHash() function instead of the
SignatureHashBitcoinCash() implementation detail.

Also enforce / check for the SIGHASH_FORKID bit being set using a
a DbgAssert(..).

* Move all SignatureHash..() to sigcommon.cpp

And also unexport the specialized functions from interpreter.h, to only
present the unified SignatureHash()-interface.

* Add notices about returning uint256(1) to SignatureHashLegacy(..)

This adds a couple warnings to the SignatureHashLegacy() function
that returning uint256(1) one to signal errors isn't the best thing
to do and should not repeated in any future hash function implementations.

* Check SignatureHash(..) return value in unit tests and benchmarks

This renames the 'one' uint256(1) value to SIGNATURE_HASH_ERROR and exports
it from interpreter.h. It further adds checks for this value where it should
not happen to the unit tests and the benchmark.

* SignatureHashBitcoinCash(): Never call without SIGHASH_FORKID

SignatureHashBitcoinCash() was never called without the fork id flag
set (as it was only called from the SignatureHash() function).

Using a special value of uint256(1) to signal an error is breaking assumptions
about ECDSA signing and this mistake should not be repeated for the Bitcoin Cash
signature function.

This removes the return-1 code-path from SignatureHashBitcoinCash
as well as adding a warning that this function is not to be used ever without
SIGHASH_FORKID set.

* CheckSig: Remove dependency on SignatureHashBitcoinCash()

Calling SignatureHashBitcoinCash() is implicit if the SIGHASH_FORKID
flag is set. This removes the explicit dependency on
SignatureHashBitcoinCash() and replaces it with a generic SignatureHash()
call.

Also removes the SignatureHashBitcoinCash() export.