Remove old unuseable code for xthins and graphene
We handle xthins and graphene block requests using get_xthin and
get_graphene. We don't allow a user to ask for these type of blocks
through a getdata request. In the very first XTHIN prototype we did
allow these requests since there was no get_xthin message type. We
instead send the bloom filter forward first and followed that with
a getdata request for a MSG_XTHINBLOCK. However, we eventually
realized this could be unreliable because of the different paths
that the filter and the getdata request would make and sometimes
the filter could arrive after the request. We then decided on
wedding the two messages together into the get_xthin.  As such this
old code path is never used and should be taken out.