Cleanup mempoolinfo which is needed to send a graphene block.
There is no need to check if the number of transactions is == -1 because
the value of nTx is a uint64_t so it can never be less than 0.

We don't need to store the nTx as a CNode value because it is never
re-used. Instead simply pass the value to SendGrapheneBlock().