Refactor bloom filter targeting for xthins and make it optional (#1183)
* Use a reference rather than copying tx and hash when doing bloom filter targeting.

* Reduce the scope of cs_main locking in bloom filter targeting

* Move code for bloom targeting into the bloom targeting code block.

This is a move only, in preparation for being able to toggle
the bloom targeting process.

* Use excessiveBlockSize as the largest block possible when targeting.

* Add command line argument use-bloom-filter-targeting

Allows us to toggle this value and leave it in the off state by default.
This feature was usuful in the past when we were running on the Bitcoin
Core chain and the mempool was often completely full. Using bloom filter
targeting in that case was very helpful in reducing the size of an
xthin, however, now that the mempool is completely cleared on the Cash
chain, we can leave this feature turned off by default.

If bloom target is not on then simply add all the mempool hashes to the filter.

* Add a node that has bloom filter targeting enabled to thinblocks.py

* Re-enable bloom filter targeting but only if mempool is too large.

Still alow the bloom targeting to be toggled and leave it on by default
but don't do the targeting unless the mempool is actually larger than EB.
That way we can still make use of the targeting if/when the mempools
get overfull.