Reduce the time it take to reach the maximum false positive rate for xthins. (#1215)
Previously we always had set this to 72 hours because on the Legacy chain
the memory pools are periodically overfull and it can take several days
after startup for the memory pools to get in sync. With Bitcoin Cash
the memory pool is cleared almost every block that is mined, so
reducing the 72 hour ramp in false positive rates,  to 12 seems like good
idea while sill leaving room in the event that we have a sudden
and likely temporary backup in the mempool.