Fix TCP drop (#1243)
If a TCP connection goes completely dead (without a close) we need to detect that by attempting a send, rather than waiting for a long time for a receive that will never happen.  Only ping out-of-time once when we decide the node is slow.  Fix some node logs to use GetLogName()