global pointer pblockdb now type CDatabaseAbstract, remove pblockundodb
Change the global pointer pblockdb to type CDatabaseAbstract so that 
during storage initialization we can assign any db class derived from 
that abstract class to be used as our block and undo storage database. 
Since Not all databases need a second pointer to a location for undo 
data, undo data pointers are now internal to the databases that do need 
them.