Prevent the re-requesting of txns
We were indicating that we had the txn when it was received however
if INV's were arriving after then we would end up re-requesting them
becuase the item had already been cleaned up from the requestmanager
queue. To fix this we just need to indicate that we received the txn
when it was either submitted to the mempool or the orphanpool. This
works because the request manager checks whether we have a txn in
either of those pools before requesting.

This is the same approach we have for blocks, where we don't indicate
we have the block until it is stored fully on disk and the blockindex
updated.