Eliminate duplicate transactions from the deferQ
Sometimes when we re-request transactions they can arrive at the same
time and appear to the deferQ as a double spend even though it is just
the same transaction. As a result they are never able to get out of the
deferQ and it builds up over time. So use a map to store them
temporarily so that duplicates are automatically removed.