Clarify forks.cpp. Use pindexTip rather than pindexPrev. (#1388)
pindexPrev was a carry over from the original fork where we used the
pindexPrev from the block to determine whether it was fork time. However
we don't just use the forking logic for blocks, it's also for different
types of transactions that may need to be validated. So, to be clear
we can just call this what it is, the current chaintip, or pindexTip.