Store transaction size within the transaction rather than using GetSerializeSize() each time. (#1398)
* Store the serialized txn size within the txn when we call GetTxSize()

This is helpful so we don't have to make repeated calls to
::GetSerializeSize().

Needed to initialize  _nTxSize(0) when constructing using a mutable
transaction.

* use GetTxSize

* Remove nTxSize from CTxMemPoolEntry

Intead of using nTxSize we just return the size of the tx directly from
the stored transaction using GetTxSize().  This saves some storage space.

* Clean up variable shadowing of nTxSize.

* fix incomplete rebase on DoS_tests.cpp

* Make nTxSize atomic

Also need to create a copy constructor because atomics are not copyable.