Make graphene use excessive-block-logic
A bunch of consensus related parameter(EB, sigops per MB, etc etc)
follow an emergent consensus logic, i.e. this was not tha case for
graphene while trying to guess estimate the number of transactions
in a block. This changeset will make it so that the constraint on the
maximum number of transactions follow take into account of
MAX_MESSAGE_SIZE_MULTIPLIER