Turn graphene on by default (#1418)
* Turn graphene on by default

* Check for HaveNodes

* Handle edge case for requesting a full block.

Although very unlikely this edge case could happen if a node
decides not to run one of either thinblocks or graphene while at
the same time doesn't have either graphene peers or thinblock peers.

So for the following cases...

Graphene disable but Thinblock enabled, however there are graphene peers
but no thinblock peers connected.

or

Graphene enabled and Thinblocks disabled, however there are thinblock peers
but not graphene peers connected.

* indent only - no semantic changes

* Prelim refactor of ipaddress() in bitcoin tests

Remove the repeating code for defining ipaddress() in the bitcoin
tests. Just have a global definition rather than having to create
multiple but slightly different definitions.

* Add request manager tests for RequestBlock()

Make RequestBlock() public

Add helper functions to util.h: SetArg(), SetBoolArg()

Make GetTimeMillis() be useable if mocktime is set.  This is so we can test
the functionlity of the thinblock and graphene timers.

* Remove redundant checked for IsGraphenBlockEnabled()