Remove a useless assert in PruneBlockIndexCandidates()
The assert at the end PruneBlockIndexCandidates() was put there
to make sure that the setBlockIndexCandidates was not empty or
abort the program otherwise.

Fact is that the code at the top of `PruneBlockIndexCandidates()`
already deal with the case of setBlockIndexCandidates being empty
by returning without raising an error. This imply that the assert
cannot be triggered.