[Windows Build Scripts] - Update toolchain and dependencies versions. (#1550)
* Include patch utility in base MinGW installation

* Update hexdump utility source to master as of 2018-12-21

A new release was tagged on 2018-12-15, however, with gcc 7.3.0 using the -Werror option causes a build failure.  This is fixed in the subsequent patch from 2018-12-21, but has not been tagged as a release yet.

* Update OpenSSL dependency to v1.0.2o

* Update libpng dependency to v1.6.36

* Update qrencode dependency to v4.0.2

* Update MinGW build toolchain to gcc v7.3.0, v5, rev0

* Update Boost dependency to v1.68.0

* Update Qt dependency to v5.7.1

* Add bitcoin-miner.exe to post-build processing

* Change default path for git install location

* Call install-utilities.sh from rebuild-bitcoin.bat

This will ensure that if the user modifies the paths to git or python in SET_ENV_VARS.bat, the executable wrappers are updated accordingly.
Without this change, the user needs to rerun config-mingw.bat to update the utility wrappers.