Prevent requesting get_blocktx, get_grblocktx and getblocktxn too far from tip
We limit how far away from the tip you can request by using the same
value DEFAULT_BLOCKS_FROM_TIP that we now use for testing if
the chain is nearly synced in IsChainNearlySyncd().  This way an
attacker can not keep our disk busy by forcing us to load blocks from
any part of the disk but will be restricted to just a few blocks which
would likely still be in cache if re-requested over and over.