Avoid clearing thintype block data in parallel.cpp (#1595)
Currently, parallel.py suffers from an intermittent failure due to a segfault in `CGrapheneBlockTx::HandleMessage`. The problem is that the size of `grapheneBlockHashes` is expected to always be at least as large as the size of `grapheneBlockTx.vMissingTx`, but this is not the case when Graphene block data is cleared while a Graphene block is in-flight. 

The fix presented here is to first eliminate the statements in parallel.cpp that clear Graphene and XThin block data, and instead clear their in-flight status in thinrelay. Additionally, this commit adds a sanity check `CGrapheneBlockTx::HandleMessage` to abort a Graphene download if ever `grapheneBlockHashes.size() < grapheneBlockTx.vMissingTx.size()`.