Seed IBLT hash functions (#1582)
Previously, in the construction of a CGrapheneSet, the ith hash function operating on input x always evaluated to the same value across all instantiations. This is undesirable because IBLT decode failures, which are pseudo-random, have the potential to occur for multiple peers when reconciling the same block.

The present PR adds the ability to pass the CIblt object an arbitrary 32-bit salt value. This salt is used to seed the IBLT hash functions. For entropy, we use the random nonce created by the Graphene block sender when seeding the SipHash used for Graphene cheap hashes.