Allow Graphene to use variant of CFastFilter (#1603)
Currently, Graphene uses CBloomFilter to store the hash values of transactions in the block. Insertion and lookup of a given transaction hash require multiple murmur hashes in order to determine which bits to set in the filter. CFastFilter is an alternative to CBloomFilter the uses the entropy in the transaction hash itself to determine filter locations. 

Unfortunately, the existing CFastFilter implementation is not appropriate for use in Graphene because 1) the filter size and number of hash functions are set at compile time, and 2) the filter size must be a power of two. In this PR, we introduce CVariableFastFilter, which is a relatively straightforward port of CFastFilter that allows for arbitrary filter sizes, configurable at runtime.

This PR introduces a command line switch: -compute-optimize-graphene (set to false by default). When the flag is set, CVariableFastFilter is used in Graphene instead of CBloomFilter. In testing, we find that compute optimized graphene offers a roughly 30% improvement in processing time when constructing a Graphene block.