Solve a few performance issues in SendMessages() (#1736)
* Swap vectors rather than copy for vBlockHashesToAnnounce

* Remove unnecessary cs_main lock

* Remove check for re-org in SendMessages()

This section, taking the cs_main lock, has a marked impact on
the performance of announcing new blocks. Surprisingly it adds
another .25 to .3 seconds of time to get the first announcements
out after processing a block. It seems from observation that as
we're trying to get these announcments out we're still receiving
headers from other peers which then also have to contend for cs_main and
this is the likely source of the contention.

It's not necessary to revert to INV if we've had a re-org as we have
to announce the blocks anyway so given the above perforance hit
it's better to remove this and get the announcements out as fast
as possible.