Do not re-request a block if it is currently being processed (#1694)
* Break up requester.Processing in two functions

Create ProcessingTxn() and ProcessingBlock(). This makes a clear
distinction between the two and removes the need for relying on
message type for determining whehter a txn or block is processing. This
is important because there are so many message types now relating to
blocks, such as graphene, thinblock, compactblock and perhaps others in
the future which we may forget to add. Furthermore, setting the block
processing flag should only be done when the block is fully reconstructed
anyway which means relying on msg_xthinblock or msg_compactblock types
is not relevant.

* Indicate we are processing the block after the block has been reconstructed.

Set the processing flag so we can prevent re-requesting this block
during the time it is being processed.

* Do not request a block if it is currently in processing