Refactor Graphene out of CNode (#1708)
* Add CGrpaheneBlock pointer to CBlockThinRelay

* Move nSizeGrapheneBlock out of CNode

* Move grapheneBlockWaitingForTxns out of CNode

* Move nLocalGrapheneBlockBytes out of CNode

* Move grapheneBlockHashes out of CNode

* Move mapGrapheneMissingTx out of CNode

* Move mapGrapheneAdditionalTxs out of CNode

* Move mapGrapheneHashOrderIndex out of CNode

* Move grapheneBlock out of CNode

* Make pointer all shared_ptrs

* Do not use references to smart pointers

* Fix bug in RequestFailOverBlock()

* Reject blocks by current size only during reconstruction

Since we only used shared pointers for transactions in the mempool, and only
begin recontruction of the block once we have all the necessary data,
it is no longer needed to be able to track bytes used for
multiple thintype blocks that are still in flight and only partially
reconstructed. This commit simplifies the entire procedure and simply
checks if the block has gone over the byte limit during reconstuction and
if so then cleanup the data and disconnet the peer.

* Re-enable excessive.py and mempool_accept.py

* Use reset() instead of setting to nullptr in SetNull()

This prevents us from ever having other threads possibly setting
our thintype objects to  nullptr in CBlockThinRelay objects. If some
other thread is still holds a shared pointer then is won't be
invalidated and the destructor won't be called.

* When setting a block to reconstruct check first if we already have one

It's possible that we could have both a thinblock and grapheneblock
from being requested for the same block hash so check first if we've
already create a shared pointer to a CBlockThinRelay object and only
if we don't have one then create a fresh instance. This way we won't
accidentally remove an instance which has an active thintype block
in use.

* Add a few DbgAsserts

Also check if ptx is null before updating block bytes

 In practice getting a null ptx could not happen because we always
 make sure we have all the transactions before we reconstruct the block
 however just in case someone makes a mistake in the future it's a good
 check to have.

* Standardize handling of smart pointer references in all thinrelay types

* Combine both thinblock and xthinblock mapMissingTx

Since we no longer use nullptr for xthinblock or thinblock which are held
in pblock then we have to either pass in the netmessage type or simply
combine both maps into one. Here, the simple approach of combining both
maps is done which prevents any issues regarding passing in an incorrect
netmessage type.