Make sure to lock cs_mapBlockIndex when accessing nStatus or nSequenceId (#1715)
* Lock nSequenceId with cs_mapBlockIndex

There were some areas that were not locked fully or properly. We have
to lock any blockindex data that is mutable. Currently only nSequenceId
and nStatus need to be locked when we read or update these values.

* When we access nStatus we must have a lock on  cs_mapBlockIndex

Generally when accessing the underlying data from a blockindex pointer
we don't have to lock anything except for the access to the map in
aquire the pointer. This is because the underlying data is immutable
with the exception of nStatus and nSequenceId.

In this commit we make sure to lock any acesses, either read or write.
to nStatus. Access to nSequenceId were dealt with in the previous
commit.

* In ProcessGetData always make sure we are using a valid blockindex

It's rather easy to make a mistake in this section of code which
could open up a remote crash attack. This could happen if a developer
failed to use the fSend flag before accessing the blockindex pointer. The fSend
flag in this section of code has been used to test whether the blockindex
is null or not, which is not a good practice.

* Fix potential deadlock

Locking order shoud be cs_main->cs_wallet->cs_mapBlockIndex but
was set to cs_main->cs_mapBlockIndex. In InvalidateBlock(), exxit
the scope where we take cs_mapBlockIndex before we DisconnectTip
which involves the wallet.

* Improve reliability of keypool.py

* Fix GetChainTips

* Reduce the scope of READLOCK(cs_mapBlockIndex)

* fHaveData as we do not really need a lock here